LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/top-level-await/sibling-imports-not-blocked.any-expected.txt


FAIL Async modules only scheduling microtasks don't block execution of sibling modules assert_equals: expected "PASS" but got "FAIL"
FAIL Async modules scheduling tasks don't block execution of sibling modules assert_equals: expected "PASS" but got "FAIL"
